build_index/requirements.txt
requests==2.25.1
tqdm==4.60.0
python-dotenv==0.17.0
